+++ src/Relude/Numeric.hs
@@ -1,4 +1,6 @@
-{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE RankNTypes          #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications    #-}
 
 {- |
 Copyright: (c) 2018 Kowainik
@@ -9,6 +11,7 @@ Maintainer: Kowainik <xrom.xkov@gmail.com>
 module Relude.Numeric
        ( module Data.Bits
        , module Data.Int
+       , module Data.Word
        , module Numeric
        , module GHC.Float
        , module GHC.Num
@@ -16,13 +19,14 @@ module Relude.Numeric
        , integerToBounded
        , integerToNatural
        )
-        where
+       where
 
 import Numeric
 
-import Data.Bits (toIntegralSized)
+import Data.Bits (toIntegralSized, xor)
 import Data.Function (($))
 import Data.Int (Int, Int16, Int32, Int64, Int8)
+import Data.Word (Word, Word16, Word32, Word64, Word8, byteSwap16, byteSwap32, byteSwap64)
 import GHC.Float (Double (..), Float (..), Floating (acos, acosh, asin, asinh, atan, atanh, cos, cosh, exp, logBase, pi, sin, sinh, sqrt, tan, tanh, (**)))
 import GHC.Num (Integer, Num (..), subtract)
 import GHC.Real (Fractional (..), Integral (..), Ratio, Rational, Real (..), RealFrac (..),
@@ -36,24 +40,37 @@ import Relude.Monad (Maybe (..))
 
 -- $setup
 -- import Relude.Monad (Maybe (..))
---
--- >>> integerToBounded 42
--- Just 42
---
--- integerToBounded :: forall a. (Integral a, Bounded a) => Integer -> Maybe a
-integerToBounded :: Int -> Maybe Int
-integerToBounded n | n < minBound = Nothing
-                   | n > maxBound = Nothing
-                   | otherwise    = Just n
 
--- >>> integerToNatural (-1)
--- Nothing
---
--- >>> integerToNatural 0
--- Nothing
---
--- >>> integerToNatural 10
--- Just 10
+{- | Transforms an integer number to a bounded integral.
+It returns `Nothing` for integers outside the bound of the return type.
+
+>>> integerToBounded @Int 42
+Just 42
+
+>>> integerToBounded @Int8 1024
+Nothing
+
+-}
+integerToBounded :: forall a. (Integral a, Bounded a) => Integer -> Maybe a
+integerToBounded n
+  | n < toInteger (minBound :: a) = Nothing
+  | n > toInteger (maxBound :: a) = Nothing
+  | otherwise                    = Just (fromIntegral n)
+
+{- | Tranforms an integer number to a natural.
+Only non-negative integers are considered natural, everything else will return `Nothing`.
+
+>>> integerToNatural (-1)
+Nothing
+
+>>> integerToNatural 0
+Nothing
+
+>>> integerToNatural 10
+Just 10
+
+-}
 integerToNatural :: Integer -> Maybe Natural
-integerToNatural n | n <= 0    = Nothing
-                   | otherwise = Just $ fromIntegral n
+integerToNatural n
+  | n <= 0    = Nothing
+  | otherwise = Just $ fromIntegral n