+++ CHANGELOG.md
@@ -3,12 +3,13 @@
 Unreleased
 ====
 
-0.5.0
+Unreleased: 0.5.0
 =====
 
 * [#125](https://github.com/kowainik/relude/issues/125):
   Moved many numerical functions and types in `Relude.Numeric`.    
-  It reexports `toIntegralSized` and exports `integerToBounded` and `integerToNatural`
+  Reexport `toIntegralSized` from `Data.Bits`.    
+  Add `integerToBounded` and `integerToNatural` in `Relude.Numeric`.
 
 0.4.0 — Nov 6, 2018
 =====